Code cleanup and refactoring.
Restructured everything to ``wwwbase/typewriter``.

``run.js`` now holds the code for both definitions and overlay and
exposes several functions in the ``window.typewriter`` namespace.

Using separate cookies for the overlay and the definitions so the
definitions will still run if the first encounter was the overlay on a
different page.